feat: creating getFavorites service. solving nft list service bug after a drop cascade on favorite table
