+++ src/user/DTO/get-favorites-dto.ts
@@ -1,0 +1,6 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class GetFavoritesDto {
+  @ApiProperty()
+  user_id: number;
+}
